refactor(fl): extract fltest framework into fl/test/ module
The portable test framework (FL_CHECK, FL_REQUIRE, FL_SUBCASE,
FL_TEST_CASE, FL_TEST_FILE) is a self-contained subsystem and now
gets its own module slot at fl/test/.

- Moves fl/fltest.{h,cpp.hpp} -> fl/test/
- Adds fl/test/_build.cpp.hpp (module unity-build aggregator)
- Adds src/fl/build/fl.test+.cpp (per-module recursive aggregator)
- Removes flat fltest includes from src/fl/_build.cpp.hpp and
  src/fl/build/fl.cpp; replaces with sub-directory aggregator pull
- Registers fl/build/fl.test+.cpp in EXPECTED_BUILD_FILES (lint check)
- Sweeps the small set of direct fl/fltest.h consumers (3 tests +
  internal self-include); most tests use tests/shared/fl_unittest.h
  which transitively includes fltest.h
- Relocates the framework's own test (tests/fl/fltest.cpp) under
  tests/fl/test/ to match TestPathStructureChecker
